+++ P2/src/parser.cpp
@@ -14,16 +14,16 @@
 using namespace std;
 
 Result ReadRectangleDimensions(fstream &input_file, 
-                               unsigned long long **rectangle_dims,
-                               unsigned long long n)
+                               uint64_t **rectangle_dims,
+                               uint64_t n)
 {
     string line;
-    unsigned long long width, height;
+    uint64_t width, height;
     /* Skip first line. */
     getline(input_file, line);
 
-    /* Read n lines and fetch two unsigned long longs from each of them. */
-    for (unsigned long long i = 0; i < n; i++) {
+    /* Read n lines and fetch two uint64_ts from each of them. */
+    for (uint64_t i = 0; i < n; i++) {
         getline(input_file, line);
         istringstream line_stream(line);
         if (line_stream >> width >> height) {
@@ -37,7 +37,7 @@ Result ReadRectangleDimensions(fstream &input_file,
 }
 
 Result ParseFile(const char *file_name, Instance &inst) {
-    unsigned long long n;
+    uint64_t n;
 
     /* Try to open file. */
     fstream input_file(file_name);
@@ -55,10 +55,10 @@ Result ParseFile(const char *file_name, Instance &inst) {
     }
 
     /* Try to read rectangles dims line by line. */
-    unsigned long long **rectangle_dims;
-    rectangle_dims = (unsigned long long**) calloc(n, sizeof(unsigned long long*));
-    for (unsigned long long i = 0; i < n; i++)
-        rectangle_dims[i] = (unsigned long long*) calloc(2, sizeof(unsigned long long));
+    uint64_t **rectangle_dims;
+    rectangle_dims = (uint64_t**) calloc(n, sizeof(uint64_t*));
+    for (uint64_t i = 0; i < n; i++)
+        rectangle_dims[i] = (uint64_t*) calloc(2, sizeof(uint64_t));
 
     if (ReadRectangleDimensions(input_file, rectangle_dims, n) == FAIL) {
         cout << "ParseFile: Failed to read rectangle dimensions" << endl;
